mqtt_grpc_proxy::mqttv5pb::mqtt_relay_service_client

Struct MqttRelayServiceClient

pub struct MqttRelayServiceClient<T> { /* private fields */ }

gRPC service for relaying MQTT packets.

Implementations

impl MqttRelayServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> MqttRelayServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> MqttRelayServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn stream_mqtt_messages( &mut self, request: impl IntoStreamingRequest<Message = MqttStreamMessage>, ) -> Result<Response<Streaming<MqttStreamMessage>>, Status>

Bidirectional streaming - one stream per MQTT connection

pub async fn relay_packet( &mut self, request: impl IntoRequest<MqttPacket>, ) -> Result<Response<RelayResponse>, Status>

Legacy unary RPCs (for backward compatibility)

pub async fn mqtt_connect( &mut self, request: impl IntoRequest<Connect>, ) -> Result<Response<Connack>, Status>

pub async fn mqtt_publish_qos1( &mut self, request: impl IntoRequest<Publish>, ) -> Result<Response<Puback>, Status>

mqtt publish RPC.

pub async fn mqtt_subscribe( &mut self, request: impl IntoRequest<Subscribe>, ) -> Result<Response<Suback>, Status>

mqtt subscribe RPC.

pub async fn mqtt_suback( &mut self, request: impl IntoRequest<Suback>, ) -> Result<Response<RelayResponse>, Status>

mqtt suback RPC - server subscription acknowledgment.

pub async fn mqtt_unsubscribe( &mut self, request: impl IntoRequest<Unsubscribe>, ) -> Result<Response<Unsuback>, Status>

mqtt unsubscribe RPC.

pub async fn mqtt_unsuback( &mut self, request: impl IntoRequest<Unsuback>, ) -> Result<Response<RelayResponse>, Status>

mqtt unsuback RPC - server unsubscription acknowledgment.

pub async fn mqtt_puback( &mut self, request: impl IntoRequest<Puback>, ) -> Result<Response<RelayResponse>, Status>

mqtt puback RPC - QoS 1 acknowledgment.

pub async fn mqtt_pubrec( &mut self, request: impl IntoRequest<Pubrec>, ) -> Result<Response<Pubrel>, Status>

mqtt pubrec RPC - QoS 2 publish received (step 1).

pub async fn mqtt_pubrel( &mut self, request: impl IntoRequest<Pubrel>, ) -> Result<Response<Pubcomp>, Status>

mqtt pubrel RPC - QoS 2 publish release (step 2).

pub async fn mqtt_pubcomp( &mut self, request: impl IntoRequest<Pubcomp>, ) -> Result<Response<RelayResponse>, Status>

mqtt pubcomp RPC - QoS 2 publish complete (step 3).

pub async fn mqtt_pingreq( &mut self, request: impl IntoRequest<Pingreq>, ) -> Result<Response<Pingresp>, Status>

mqtt pingreq RPC - client ping request.

pub async fn mqtt_disconnect( &mut self, request: impl IntoRequest<Disconnect>, ) -> Result<Response<RelayResponse>, Status>

mqtt disconnect RPC - graceful disconnection.

pub async fn mqtt_auth( &mut self, request: impl IntoRequest<Auth>, ) -> Result<Response<Auth>, Status>

mqtt auth RPC - authentication exchange (MQTT 5.0).