mqtt_grpc_proxy::mqttv5pb::property

Enum PropertyType

pub enum PropertyType {

    PayloadFormatIndicator(bool),
    MessageExpiryInterval(u32),
    ContentType(String),
    ResponseTopic(String),
    CorrelationData(Vec<u8>),
    SubscriptionIdentifier(u32),
    SessionExpiryInterval(u32),
    AssignedClientIdentifier(String),
    ServerKeepAlive(u32),
    AuthenticationMethod(String),
    AuthenticationData(Vec<u8>),
    RequestProblemInformation(bool),
    WillDelayInterval(u32),
    RequestResponseInformation(bool),
    ResponseInformation(String),
    ServerReference(String),
    ReasonString(String),
    ReceiveMaximum(u32),
    TopicAliasMaximum(u32),
    TopicAlias(u32),
    MaximumQos(u32),
    RetainAvailable(bool),
    UserProperty(UserProperty),
    MaximumPacketSize(u32),
    WildcardSubscriptionAvailable(bool),
    SubscriptionIdentifiersAvailable(bool),
    SharedSubscriptionAvailable(bool),

}

Variants

PayloadFormatIndicator(bool)

MessageExpiryInterval(u32)

ContentType(String)

ResponseTopic(String)

CorrelationData(Vec<u8>)

SubscriptionIdentifier(u32)

SessionExpiryInterval(u32)

AssignedClientIdentifier(String)

ServerKeepAlive(u32)

AuthenticationMethod(String)

AuthenticationData(Vec<u8>)

RequestProblemInformation(bool)

WillDelayInterval(u32)

RequestResponseInformation(bool)

ResponseInformation(String)

ServerReference(String)

ReasonString(String)

ReceiveMaximum(u32)

TopicAliasMaximum(u32)

TopicAlias(u32)

MaximumQos(u32)

RetainAvailable(bool)

UserProperty(UserProperty)

MaximumPacketSize(u32)

WildcardSubscriptionAvailable(bool)

SubscriptionIdentifiersAvailable(bool)

SharedSubscriptionAvailable(bool)

Implementations

impl PropertyType

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<PropertyType>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.